Use interfaces in the Transcript class. Remove moved files.
